Updated keymap.c - modification to caps lock key and settings layer
What was formerly the CAPS (caps lock) key -- is now ESC.
    However, if this key is held down, it becomes Fn. (enter the function layer).
    In addition, the Fn key on R0 has been removed and replaced
        with Right Alt (on KM_MACOS)
        or Right GUI (on KM_WIN)

    What was formerly ESC is now CAPS.

You will now have regular 7u bottom row for both KM_MACOS and KM_WIN
(with respect to RAlt/RGUI being available, as opposed to Fn taking its place).
The Caps Lock key is better served as Esc when tapped, but Fn when held (since Fn is typically held anyway!)

If you need access to a regular Esc key (say, if Esc needs to be held down),
Press and hold the "new" Esc key to activate the Fn layer, which maps Esc to where it was before.

See keymap.c for the layout.