Updated Caps Lock functionality (now Esc) for KM_MACOS and KM_WIN
The Caps Lock key has been remapped to be Esc --
but when this key is held, it becomes Fn.

The Fn key in R0 has been removed, and is now RAlt/RGUI
(dependent if in macOS layer or Windows layer).

The Esc key has been remapped to be Caps Lock.
When a "held" Esc key is required, invoke the Fn layer
by holding the "new" Esc key, then press and hold the "old" Esc key.
The Fn layer has Esc mapped to its "traditional" location.