Added Apple fn key functionality to unloved_bastard
In keymap KM_MACOS, the keycode KC_INS (insert key) was replaced with KC_APFN (Apple fn key).

In "macOS-centric" keyboards, this is the norm.
The KC_APFN keycode functionality was added in a previous commit on this branch -- I had retrieved a diff from issue 2179 (written by user fauxpark), and patched it in to this local copy of qmk_firmware.

in the KM_FUNC keymap, the key where the Apple fn key would be is no longer transparent -- but rather, it is now KC_INS. This will allow a user to have temporary access to a traditional Insert key while their primary layer is KM_MACOS.

The KM_WIN keymap remains unchanged.