Updated led_set_user function for Filco-style cases
The Unloved Bastard (as does the CoolerMaster MasterKeys S)
has three LEDs on the controller - but these LEDs get in the way
of the YMDK Filco-style TKL case.

I have rewired two of the three LEDs (the left and middle ones)
so that they are wired away from the controller -- so that the
LEDs can reside where they would on a Filco case.

The third LED (right most one) isn't needed, or rather,
Filco TKLs don't have three LEDs -- so it has been permanently removed.

In led_set_user, I've modified the code such that
the middle LED (now the right-most LED on the case)
is always on -- this indicates that the keyboard is operational.
The left LED retains its caps lock indicator function --
but I may repurpose this LED if and when I remove caps lock
from all key mappings.